backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
yfinance>=0.2.40
nsetools==2.0.1
nsepython==2.97
google-generativeai
requests-cache==1.2.0